refactor: move tests to parallel test directory
- Create test/ directory structure mirroring src/
- Move all test files from src/ to test/
- Update all import paths in test files to reference ../src/
- Update jest.config.mjs to use new test directory location
- All tests passing (24 suites, 367 tests passed)

Resolves #187
